+++ template/addons.php
@@ -4,10 +4,10 @@
 <div class="addons-tab content-tab">
 	<div class="addons-content">
 		<?php if( ! empty( $licence ) ) : ?>
-			<p><?php _e( 'Fetching addon details, please wait...', 'wp-migrate-db-pro' ); ?></p>
+			<p><?php _e( 'Fetching addon details, please wait...', 'wp-migrate-db' ); ?></p>
 		<?php else : ?>
-			<p><?php _e( 'We couldn\'t find your license information. Please switch to the settings tab and enter your license.', 'wp-migrate-db-pro' ); ?></p>
-			<p><?php _e( 'Once completed, you may visit this tab to view the available addons.', 'wp-migrate-db-pro' ); ?></p>
+			<p><?php _e( 'We couldn\'t find your license information. Please switch to the settings tab and enter your license.', 'wp-migrate-db' ); ?></p>
+			<p><?php _e( 'Once completed, you may visit this tab to view the available addons.', 'wp-migrate-db' ); ?></p>
 		<?php endif; ?>
 	</div>
 </div>
