+++ template/common/debug-info.php
@@ -1,0 +1,5 @@
+<div class="debug">
+	<h3><?php _e( 'Diagnostic Info &amp; Error Log', 'wp-migrate-db' ); ?></h3>
+	<textarea class="debug-log-textarea" autocomplete="off" readonly></textarea>
+	<a class="button clear-log js-action-link"><?php _e( 'Clear Error Log', 'wp-migrate-db' ); ?></a>
+</div>
