+++ template/migrate-progress.php
@@ -1,11 +1,11 @@
 <div class="progress-content">
 	<span class="close-progress-content close-progress-content-button">&times;</span>
 	<div>
-		<h2 class="progress-title"><?php _e( 'Please wait while migration is running...', 'wp-migrate-db-pro' ); ?></h2>
+		<h2 class="progress-title"><?php _e( 'Please wait while migration is running...', 'wp-migrate-db' ); ?></h2>
 	</div>
 	<div class="progress-info-wrapper clearfix">
-		<div class="progress-text"><?php _e( 'Establishing Connection', 'wp-migrate-db-pro' ); ?></div>
-		<span class="timer"><?php echo __( 'Time Elapsed:', 'wp-migrate-db-pro' ) . ' 00:00:00'; ?></span>
+		<div class="progress-text"><?php _e( 'Establishing Connection', 'wp-migrate-db' ); ?></div>
+		<span class="timer"><?php echo __( 'Time Elapsed:', 'wp-migrate-db' ) . ' 00:00:00'; ?></span>
 	</div>
 	<div class="clearfix"></div>
 	<div class="progress-bar-wrapper">
@@ -16,8 +16,10 @@
 	</div>
 
 	<div class="migration-controls">
-		<span class="pause-resume button"><?php _e( 'Pause', 'wp-migrate-db-pro' ); ?></span>
-		<span class="cancel button"><?php _e( 'Cancel', 'wp-migrate-db-pro' ); ?></span>
+		<span class="pause-resume button"><?php _e( 'Pause', 'wp-migrate-db' ); ?></span>
+		<span class="cancel button"><?php _e( 'Cancel', 'wp-migrate-db' ); ?></span>
 	</div>
+
+	<?php $this->template_part( array( 'progress_upgrade' ) ); ?>
 
 </div> <!-- end .progress-content -->
